+++ skbold/utils/convert2mni.py
@@ -65,7 +65,9 @@ def convert2mni(file2transform, reg_dir, out_dir=None, interpolation='trilinear'
         apply_xfm.inputs.apply_xfm = True
         apply_xfm.run()
 
-        os.remove(out_matrix_file)
+        if op.exists(out_matrix_file):
+            os.remove(out_matrix_file)
+
         out_all.append(out_file)
         out_name = None
